Split space into seperate files
