+++ mesa/space/__init__.py
@@ -1,0 +1,4 @@
+from .hexgrid import *
+from .grid import *
+from .network import *
+from .continuous_space import *
