+++ mesa/space/grid.py
@@ -19,7 +19,7 @@ import itertools
 import numpy as np
 
 from typing import Iterable, Iterator, List, Optional, Set, Tuple, Union
-from .agent import Agent
+from mesa.agent import Agent
 
 Coordinate = Tuple[int, int]
 GridContent = List[Agent]
@@ -247,8 +247,8 @@ class MultiGrid:
 
         """
         neighborhood = self.get_neighborhood(pos, moore, include_center, radius)
-        return self.get_contents(neighborhood)
-        # return list(itertools.chain.from_iterable(neighbors))
+        neighbors = self.get_contents(neighborhood)
+        return list(itertools.chain.from_iterable(neighbors))
 
     def neighbor_iter(
         self, pos: Coordinate, moore: bool = True
@@ -384,6 +384,36 @@ class SingleGrid(MultiGrid):
         else:
             raise Exception("Cell not empty")
 
+    def get_neighbors(
+        self,
+        pos: Coordinate,
+        moore: bool = True,
+        include_center: bool = False,
+        radius: int = 1,
+    ) -> List[Agent]:
+        """ Return a list of neighbors to a certain point.
+
+        Args:
+            pos: Coordinate tuple for the neighborhood to get.
+            moore: If True, return Moore neighborhood
+                    (including diagonals)
+                   If False, return Von Neumann neighborhood
+                     (exclude diagonals)
+            include_center: If True, return the (x, y) cell as well.
+                            Otherwise,
+                            return surrounding cells only.
+            radius: radius, in cells, of neighborhood to get.
+
+        Returns:
+            A list of non-None objects in the given neighborhood;
+            at most 9 if Moore, 5 if Von-Neumann
+            (8 and 4 if not including the center).
+
+        """
+        neighborhood = self.get_neighborhood(pos, moore, include_center, radius)
+        return self.get_contents(neighborhood)
+        # return list(itertools.chain.from_iterable(neighbors))
+
 
 class Grid(SingleGrid):
     """ Grid where each cell can contain more than one object.
@@ -404,377 +434,3 @@ class Grid(SingleGrid):
     Methods:
         get_neighbors: Returns the objects surrounding a given cell.
     """
-
-
-class HexGrid(SingleGrid):
-    """ Hexagonal Grid: Extends Grid to handle hexagonal neighbors.
-
-    Functions according to odd-q rules.
-    See http://www.redblobgames.com/grids/hexagons/#coordinates for more.
-
-    Properties:
-        width, height: The grid's width and height.
-        torus: Boolean which determines whether to treat the grid as a torus.
-
-    Methods:
-        get_neighbors: Returns the objects surrounding a given cell.
-        get_neighborhood: Returns the cells surrounding a given cell.
-        neighbor_iter: Iterates over position neightbors.
-        iter_neighborhood: Returns an iterator over cell coordinates that are
-            in the neighborhood of a certain point.
-
-    """
-
-    def iter_neighborhood(self, pos, include_center=False, radius=1):
-        """ Return an iterator over cell coordinates that are in the
-        neighborhood of a certain point.
-
-        Args:
-            pos: Coordinate tuple for the neighborhood to get.
-            include_center: If True, return the (x, y) cell as well.
-                            Otherwise, return surrounding cells only.
-            radius: radius, in cells, of neighborhood to get.
-
-        Returns:
-            A list of coordinate tuples representing the neighborhood. For
-            example with radius 1, it will return list with number of elements
-            equals at most 9 (8) if Moore, 5 (4) if Von Neumann (if not
-            including the center).
-
-        """
-
-        def torus_adj_2d(pos):
-            return (pos[0] % self.width, pos[1] % self.height)
-
-        coordinates = set()
-
-        def find_neighbors(pos, radius):
-            x, y = pos
-
-            """
-            Both: (0,-), (0,+)
-
-            Even: (-,+), (-,0), (+,+), (+,0)
-            Odd:  (-,0), (-,-), (+,0), (+,-)
-            """
-            adjacent = [(x, y - 1), (x, y + 1)]
-
-            if include_center:
-                adjacent.append(pos)
-
-            if x % 2 == 0:
-                adjacent += [(x - 1, y + 1), (x - 1, y), (x + 1, y + 1), (x + 1, y)]
-            else:
-                adjacent += [(x - 1, y), (x - 1, y - 1), (x + 1, y), (x + 1, y - 1)]
-
-            if self.torus is False:
-                adjacent = list(
-                    filter(lambda coords: not self.out_of_bounds(coords), adjacent)
-                )
-            else:
-                adjacent = [torus_adj_2d(coord) for coord in adjacent]
-
-            coordinates.update(adjacent)
-
-            if radius > 1:
-                [find_neighbors(coords, radius - 1) for coords in adjacent]
-
-        find_neighbors(pos, radius)
-
-        if not include_center and pos in coordinates:
-            coordinates.remove(pos)
-
-        for i in coordinates:
-            yield i
-
-    def neighbor_iter(self, pos):
-        """ Iterate over position neighbors.
-
-        Args:
-            pos: (x,y) coords tuple for the position to get the neighbors of.
-
-        """
-        neighborhood = self.iter_neighborhood(pos)
-        return self.iter_cell_list_contents(neighborhood)
-
-    def get_neighborhood(self, pos, include_center=False, radius=1):
-        """ Return a list of cells that are in the neighborhood of a
-        certain point.
-
-        Args:
-            pos: Coordinate tuple for the neighborhood to get.
-            include_center: If True, return the (x, y) cell as well.
-                            Otherwise, return surrounding cells only.
-            radius: radius, in cells, of neighborhood to get.
-
-        Returns:
-            A list of coordinate tuples representing the neighborhood;
-            With radius 1
-
-        """
-        return list(self.iter_neighborhood(pos, include_center, radius))
-
-    def iter_neighbors(self, pos, include_center=False, radius=1):
-        """ Return an iterator over neighbors to a certain point.
-
-        Args:
-            pos: Coordinates for the neighborhood to get.
-            include_center: If True, return the (x, y) cell as well.
-                            Otherwise,
-                            return surrounding cells only.
-            radius: radius, in cells, of neighborhood to get.
-
-        Returns:
-            An iterator of non-None objects in the given neighborhood
-
-        """
-        neighborhood = self.iter_neighborhood(pos, include_center, radius)
-        return self.iter_cell_list_contents(neighborhood)
-
-    def get_neighbors(self, pos, include_center=False, radius=1):
-        """ Return a list of neighbors to a certain point.
-
-        Args:
-            pos: Coordinate tuple for the neighborhood to get.
-            include_center: If True, return the (x, y) cell as well.
-                            Otherwise,
-                            return surrounding cells only.
-            radius: radius, in cells, of neighborhood to get.
-
-        Returns:
-            A list of non-None objects in the given neighborhood
-
-        """
-        return list(self.iter_neighbors(pos, include_center, radius))
-
-
-class ContinuousSpace:
-    """ Continuous space where each agent can have an arbitrary position.
-
-    Assumes that all agents are point objects, and have a pos property storing
-    their position as an (x, y) tuple. This class uses a numpy array internally
-    to store agent objects, to speed up neighborhood lookups.
-
-    """
-
-    _grid = None
-
-    def __init__(self, x_max, y_max, torus, x_min=0, y_min=0):
-        """ Create a new continuous space.
-
-        Args:
-            x_max, y_max: Maximum x and y coordinates for the space.
-            torus: Boolean for whether the edges loop around.
-            x_min, y_min: (default 0) If provided, set the minimum x and y
-                          coordinates for the space. Below them, values loop to
-                          the other edge (if torus=True) or raise an exception.
-
-        """
-        self.x_min = x_min
-        self.x_max = x_max
-        self.width = x_max - x_min
-        self.y_min = y_min
-        self.y_max = y_max
-        self.height = y_max - y_min
-        self.center = np.array(((x_max + x_min) / 2, (y_max + y_min) / 2))
-        self.size = np.array((self.width, self.height))
-        self.torus = torus
-
-        self._agent_points = None
-        self._index_to_agent = {}
-        self._agent_to_index = {}
-
-    def place_agent(self, agent, pos):
-        """ Place a new agent in the space.
-
-        Args:
-            agent: Agent object to place.
-            pos: Coordinate tuple for where to place the agent.
-
-        """
-        pos = self.torus_adj(pos)
-        if self._agent_points is None:
-            self._agent_points = np.array([pos])
-        else:
-            self._agent_points = np.append(self._agent_points, np.array([pos]), axis=0)
-        self._index_to_agent[self._agent_points.shape[0] - 1] = agent
-        self._agent_to_index[agent] = self._agent_points.shape[0] - 1
-        agent.pos = pos
-
-    def move_agent(self, agent, pos):
-        """ Move an agent from its current position to a new position.
-
-        Args:
-            agent: The agent object to move.
-            pos: Coordinate tuple to move the agent to.
-
-        """
-        pos = self.torus_adj(pos)
-        idx = self._agent_to_index[agent]
-        self._agent_points[idx, 0] = pos[0]
-        self._agent_points[idx, 1] = pos[1]
-        agent.pos = pos
-
-    def remove_agent(self, agent):
-        """ Remove an agent from the simulation.
-
-        Args:
-            agent: The agent object to remove
-            """
-        if agent not in self._agent_to_index:
-            raise Exception("Agent does not exist in the space")
-        idx = self._agent_to_index[agent]
-        del self._agent_to_index[agent]
-        max_idx = max(self._index_to_agent.keys())
-        # Delete the agent's position and decrement the index/agent mapping
-        self._agent_points = np.delete(self._agent_points, idx, axis=0)
-        for a, index in self._agent_to_index.items():
-            if index > idx:
-                self._agent_to_index[a] = index - 1
-                self._index_to_agent[index - 1] = a
-        # The largest index is now redundant
-        del self._index_to_agent[max_idx]
-        agent.pos = None
-
-    def get_neighbors(self, pos, radius, include_center=True):
-        """ Get all objects within a certain radius.
-
-        Args:
-            pos: (x,y) coordinate tuple to center the search at.
-            radius: Get all the objects within this distance of the center.
-            include_center: If True, include an object at the *exact* provided
-                            coordinates. i.e. if you are searching for the
-                            neighbors of a given agent, True will include that
-                            agent in the results.
-
-        """
-        deltas = np.abs(self._agent_points - np.array(pos))
-        if self.torus:
-            deltas = np.minimum(deltas, self.size - deltas)
-        dists = deltas[:, 0] ** 2 + deltas[:, 1] ** 2
-
-        idxs, = np.where(dists <= radius ** 2)
-        neighbors = [
-            self._index_to_agent[x] for x in idxs if include_center or dists[x] > 0
-        ]
-        return neighbors
-
-    def get_heading(self, pos_1, pos_2):
-        """ Get the heading angle between two points, accounting for toroidal space.
-
-        Args:
-            pos_1, pos_2: Coordinate tuples for both points.
-        """
-        one = np.array(pos_1)
-        two = np.array(pos_2)
-        if self.torus:
-            one = (one - self.center) % self.size
-            two = (two - self.center) % self.size
-        heading = two - one
-        if isinstance(pos_1, tuple):
-            heading = tuple(heading)
-        return heading
-
-    def get_distance(self, pos_1, pos_2):
-        """ Get the distance between two point, accounting for toroidal space.
-
-        Args:
-            pos_1, pos_2: Coordinate tuples for both points.
-
-        """
-        x1, y1 = pos_1
-        x2, y2 = pos_2
-
-        dx = np.abs(x1 - x2)
-        dy = np.abs(y1 - y2)
-        if self.torus:
-            dx = min(dx, self.width - dx)
-            dy = min(dy, self.height - dy)
-        return np.sqrt(dx * dx + dy * dy)
-
-    def torus_adj(self, pos):
-        """ Adjust coordinates to handle torus looping.
-
-        If the coordinate is out-of-bounds and the space is toroidal, return
-        the corresponding point within the space. If the space is not toroidal,
-        raise an exception.
-
-        Args:
-            pos: Coordinate tuple to convert.
-
-        """
-        if not self.out_of_bounds(pos):
-            return pos
-        elif not self.torus:
-            raise Exception("Point out of bounds, and space non-toroidal.")
-        else:
-            x = self.x_min + ((pos[0] - self.x_min) % self.width)
-            y = self.y_min + ((pos[1] - self.y_min) % self.height)
-            if isinstance(pos, tuple):
-                return (x, y)
-            else:
-                return np.array((x, y))
-
-    def out_of_bounds(self, pos):
-        """ Check if a point is out of bounds. """
-        x, y = pos
-        return x < self.x_min or x >= self.x_max or y < self.y_min or y >= self.y_max
-
-
-class NetworkGrid:
-    """ Network Grid where each node contains zero or more agents. """
-
-    def __init__(self, G):
-        self.G = G
-        for node_id in self.G.nodes:
-            G.nodes[node_id]["agent"] = list()
-
-    def place_agent(self, agent, node_id):
-        """ Place a agent in a node. """
-
-        self._place_agent(agent, node_id)
-        agent.pos = node_id
-
-    def get_neighbors(self, node_id, include_center=False):
-        """ Get all adjacent nodes """
-
-        neighbors = list(self.G.neighbors(node_id))
-        if include_center:
-            neighbors.append(node_id)
-
-        return neighbors
-
-    def move_agent(self, agent, node_id):
-        """ Move an agent from its current node to a new node. """
-
-        self._remove_agent(agent, agent.pos)
-        self._place_agent(agent, node_id)
-        agent.pos = node_id
-
-    def _place_agent(self, agent, node_id):
-        """ Place the agent at the correct node. """
-
-        self.G.nodes[node_id]["agent"].append(agent)
-
-    def _remove_agent(self, agent, node_id):
-        """ Remove an agent from a node. """
-
-        self.G.nodes[node_id]["agent"].remove(agent)
-
-    def is_cell_empty(self, node_id):
-        """ Returns a bool of the contents of a cell. """
-        return not self.G.nodes[node_id]["agent"]
-
-    def get_cell_list_contents(self, cell_list):
-        return list(self.iter_cell_list_contents(cell_list))
-
-    def get_all_cell_contents(self):
-        return list(self.iter_cell_list_contents(self.G))
-
-    def iter_cell_list_contents(self, cell_list):
-        list_of_lists = [
-            self.G.nodes[node_id]["agent"]
-            for node_id in cell_list
-            if not self.is_cell_empty(node_id)
-        ]
-        return [item for sublist in list_of_lists for item in sublist]
